backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "numpy<2",
    "opencv-python<4.13",
    "pillow>=12.2.0",
    "segment-anything",
    "torch>=2.1,<2.3",
    "torchvision>=0.16,<0.18",
    "ultralytics>=8.4.41",
]

[tool.uv.sources]
segment-anything = { git = "https://github.com/facebookresearch/segment-anything.git" }
